% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{aape}
\alias{aape}
\title{Arctangent Absolute Percentage Error}
\usage{
aape(a, f)
}
\arguments{
\item{a}{\code{\link[=numeric]{numeric()}} \code{\link[=vector]{vector()}} of actual values}

\item{f}{\code{\link[=numeric]{numeric()}} \code{\link[=vector]{vector()}} of forecast values}
}
\value{
\code{\link[=numeric]{numeric()}} \code{\link[=vector]{vector()}} of arctangent absolute percentage errors
}
\description{
Arctangent Absolute Percentage Error
}
\examples{
a <- seq(0, 10, 0.1)
f <- seq(10, 0, -0.1)
m <- aape(a, f)
plot(m, type = "l")

}
\references{
Kim and Kim 2016 doi.org/10.1016/j.ijforecast.2015.12.003
}
