% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{ape}
\alias{ape}
\title{Absolute Percentage Error}
\usage{
ape(a, f)
}
\arguments{
\item{a}{\code{\link[=numeric]{numeric()}} \code{\link[=vector]{vector()}} of actual values}

\item{f}{\code{\link[=numeric]{numeric()}} \code{\link[=vector]{vector()}} of forecast values}
}
\value{
\code{\link[=numeric]{numeric()}} \code{\link[=vector]{vector()}} of absolute percentage errors
}
\description{
Absolute Percentage Error
}
\examples{
a <- seq(0, 10, 0.1)
f <- seq(10, 0, -0.1)
m <- ape(a, f)
plot(m, type = "l")

}
