% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{arctan_abs_pct_error}
\alias{arctan_abs_pct_error}
\title{Arctangent Absolute Percentage Error}
\usage{
arctan_abs_pct_error(x, y)
}
\arguments{
\item{x}{\code{\link[=numeric]{numeric()}} \code{\link[=vector]{vector()}} actual values}

\item{y}{\code{\link[=numeric]{numeric()}} \code{\link[=vector]{vector()}} forecast values}
}
\value{
\code{\link[=numeric]{numeric()}} \code{\link[=vector]{vector()}} arctangent absolute percentage errors
}
\description{
Arctangent Absolute Percentage Error
}
\examples{
x <- seq(0, 10, 0.1)
y <- seq(10, 0, -0.1)
m <- arctan_abs_pct_error(x, y)
plot(m, type = "l")

}
\references{
Kim and Kim 2016 doi.org/10.1016/j.ijforecast.2015.12.003
}
