% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{chilko}
\alias{chilko}
\title{Chilko Sockeye Salmon Abundance Data}
\format{
\code{\link[tibble:tibble]{tibble::tibble()}}
}
\usage{
chilko
}
\description{
Spawner and Recruitment data
}
\keyword{datasets}
