% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{fit}
\alias{fit}
\title{Fit A Larkin or Ricker Model Via CmdStanR}
\usage{
fit(
  data,
  model = "larkin",
  chains = 1,
  step_size = 0.01,
  iter_warmup = 250,
  iter_sampling = 750,
  ...
)
}
\arguments{
\item{data}{\code{\link[=list]{list()}}}

\item{model}{\code{\link[=character]{character()}} either \code{"larkin"} or \code{"ricker"}}

\item{chains}{\code{\link[=integer]{integer()}} number of chains}

\item{step_size}{\code{\link[=integer]{integer()}} initial step size}

\item{iter_warmup}{\code{\link[=integer]{integer()}} number of warmup iterations}

\item{iter_sampling}{\code{\link[=integer]{integer()}} number of sampling iterations}

\item{...}{additional arguments to pass to \code{$sample()} method}
}
\value{
\code{\link[=list]{list()}}
}
\description{
Fit A Larkin or Ricker Model Via CmdStanR
}
