% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{forecast}
\alias{forecast}
\title{Forecast From A Ricker Or Larkin Model Using Previous Values Only}
\usage{
forecast(
  inds,
  data,
  model = "larkin",
  window = 10,
  cores = NULL,
  chains = 1,
  step_size = 0.01,
  iter_warmup = 250,
  iter_sampling = 750,
  ...
)
}
\arguments{
\item{inds}{\code{\link[=numeric]{numeric()}} \code{\link[=vector]{vector()}} indexes of values to forecast}

\item{data}{\code{\link[=list]{list()}}}

\item{model}{\code{\link[=character]{character()}} either \code{"larkin"} or \code{"ricker"}}

\item{window}{\code{\link[=integer]{integer()}} length of the buffer between training and forecast}

\item{cores}{\code{\link[=numeric]{numeric()}} number of cores for parallel processing}

\item{chains}{\code{\link[=integer]{integer()}} number of chains}

\item{step_size}{\code{\link[=integer]{integer()}} initial step size}

\item{iter_warmup}{\code{\link[=integer]{integer()}} number of warmup iterations}

\item{iter_sampling}{\code{\link[=integer]{integer()}} number of sampling iterations}

\item{...}{additional arguments to pass to \code{$sample()} method}
}
\value{
\code{\link[tibble:tibble]{tibble::tibble()}}
}
\description{
Forecast From A Ricker Or Larkin Model Using Previous Values Only
}
