% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{forecast_single_value}
\alias{forecast_single_value}
\title{Forecast Single Value Using Previous Values Only}
\usage{
forecast_single_value(
  index,
  data,
  chains = 1,
  step_size = 0.01,
  iter_warmup = 250,
  iter_sampling = 750,
  ...
)
}
\arguments{
\item{index}{\code{\link[=numeric]{numeric()}}}

\item{data}{\code{\link[=list]{list()}}}

\item{chains}{\code{\link[=integer]{integer()}} number of chains}

\item{step_size}{\code{\link[=integer]{integer()}} initial step size}

\item{iter_warmup}{\code{\link[=integer]{integer()}} number of warmup iterations}

\item{iter_sampling}{\code{\link[=integer]{integer()}} number of sampling iterations}

\item{...}{additional arguments to pass to \code{$sample()} method}
}
\value{
\code{\link[tibble:tibble]{tibble::tibble()}}
}
\description{
Forecast Single Value Using Previous Values Only
}
