% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{harvest_control_rule}
\alias{harvest_control_rule}
\title{Harvest Control Rule}
\usage{
harvest_control_rule(x, lrp, usr, h_min = 0, h_max = 0.1)
}
\arguments{
\item{x}{\code{\link[=numeric]{numeric()}}}

\item{lrp}{\code{\link[=numeric]{numeric()}}}

\item{usr}{\code{\link[=numeric]{numeric()}}}

\item{h_min}{\code{\link[=numeric]{numeric()}}}

\item{h_max}{\code{\link[=numeric]{numeric()}}}
}
\value{
\code{\link[=numeric]{numeric()}}
}
\description{
Harvest Control Rule
}
\examples{
x <- seq(0, 2, 0.001)
lrp <- 0.001
usr <- 1
h_max <- 0.9
h <- harvest_control_rule(x = x, lrp = lrp, usr = usr, h_max = h_max)

}
