% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{maape}
\alias{maape}
\title{Mean Arctangent Absolute Percentage Error}
\usage{
maape(a, f)
}
\arguments{
\item{a}{\code{\link[=numeric]{numeric()}} \code{\link[=vector]{vector()}} of actual values}

\item{f}{\code{\link[=numeric]{numeric()}} \code{\link[=vector]{vector()}} of forecast values}
}
\value{
\code{\link[=numeric]{numeric()}} mean arctangent absolute percentage error
}
\description{
Mean Arctangent Absolute Percentage Error
}
\examples{
maape(1:10, 10:1)

}
\references{
Kim and Kim 2016 doi.org/10.1016/j.ijforecast.2015.12.003
}
