% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{mae}
\alias{mae}
\title{Mean Absolute Error}
\usage{
mae(a, f)
}
\arguments{
\item{a}{\code{\link[=numeric]{numeric()}} \code{\link[=vector]{vector()}} of actual values}

\item{f}{\code{\link[=numeric]{numeric()}} \code{\link[=vector]{vector()}} of forecast values}
}
\value{
\code{\link[=numeric]{numeric()}} mean absolute error
}
\description{
Mean Absolute Error
}
\examples{
mae(1:10, 10:1)

}
