% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{mamse}
\alias{mamse}
\title{Mean Absolute Mean-Scaled Error}
\usage{
mamse(x, y)
}
\arguments{
\item{x}{\link[=vector]{numeric()} observed values}

\item{y}{\link[=vector]{numeric()} forecast values}
}
\value{
\code{\link[=numeric]{numeric()}} scalar
}
\description{
Mean Absolute Mean-Scaled Error
}
\examples{
x <- rnorm(100, 0, 1)
y1 <- x
mamse(x, y1)

y2 <- rep(mean(x), length(x))
mamse(x, y2)

mamse(c(x, NA), c(y1, NA))

}
