% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{mape}
\alias{mape}
\title{Mean Absolute Percentage Error}
\usage{
mape(x, y)
}
\arguments{
\item{x}{\link[=vector]{numeric()} observed values}

\item{y}{\link[=vector]{numeric()} forecast values}
}
\value{
\code{\link[=numeric]{numeric()}} mean absolute percentage error
}
\description{
Mean Absolute Percentage Error
}
\examples{
mape(1:10, 10:1)

}
