% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{mase}
\alias{mase}
\title{Mean Absolute Lag-1-Scaled Error}
\usage{
mase(x, y)
}
\arguments{
\item{x}{\link[=vector]{numeric()} observed values}

\item{y}{\link[=vector]{numeric()} forecast values}
}
\value{
\code{\link[=numeric]{numeric()}} scalar
}
\description{
Mean Absolute Lag-1-Scaled Error
}
\examples{
x <- rnorm(100, 0, 1)
y1 <- x
mase(x, y1)

y2 <- dplyr::lag(x)
mase(x, y2)

mase(c(x, NA), c(y1, NA))

}
