% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{matric}
\alias{matric}
\title{Metric Wrapper}
\usage{
matric(x, fun, ...)
}
\arguments{
\item{x}{\link[=matrix]{numeric()} of two columns}

\item{fun}{\code{\link[=function]{function()}} taking two \link[=vector]{numeric()} arguments and
returning one scalar \code{\link[=numeric]{numeric()}} value}

\item{...}{additional arguments to pass to \code{fun()}}
}
\value{
\code{\link[=numeric]{numeric()}} scalar
}
\description{
Metric Wrapper
}
\examples{
x <- rnorm(100, 0, 1)
x1 <- matrix(c(x, x), ncol = 2)
matric(x1, mase)

x2 <- rbind(x1, matrix(c(NA_real_, NA_real_), ncol = 2))
matric(x2, mase)

}
