% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{mean_or_na}
\alias{mean_or_na}
\title{Arithmetic Mean Or NA}
\usage{
mean_or_na(x, max_na = Inf)
}
\arguments{
\item{x}{\code{\link[=numeric]{numeric()}} \code{\link[=vector]{vector()}}}

\item{max_na}{\code{\link[=integer]{integer()}} scalar}
}
\value{
\code{\link[=numeric]{numeric()}} scalar or \code{NA_real_}
}
\description{
Arithmetic Mean Or NA
}
\examples{
# Ends in NA
x <- c(1:9, NA)
mean_or_na(x)

# Too many NA values
x <- c(1:3, NA, NA, NA, 7:10)
mean_or_na(x, 2)

# Mean
x <- 1:10
mean_or_na(x)

}
