% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{mpe}
\alias{mpe}
\title{Mean Percentage Error}
\usage{
mpe(a, f)
}
\arguments{
\item{a}{\code{\link[=numeric]{numeric()}} \code{\link[=vector]{vector()}} of actual values}

\item{f}{\code{\link[=numeric]{numeric()}} \code{\link[=vector]{vector()}} of forecast values}
}
\value{
\code{\link[=numeric]{numeric()}} mean percentage error
}
\description{
Mean Percentage Error
}
\examples{
mpe(1:10, 10:1)

}
