% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{mre}
\alias{mre}
\title{Mean Raw Error}
\usage{
mre(a, f)
}
\arguments{
\item{a}{\code{\link[=numeric]{numeric()}} \code{\link[=vector]{vector()}} of actual values}

\item{f}{\code{\link[=numeric]{numeric()}} \code{\link[=vector]{vector()}} of forecast values}
}
\value{
\code{\link[=numeric]{numeric()}} mean raw error
}
\description{
Mean Raw Error
}
\examples{
mre(1:10, 10:1)

}
