% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{pct_error}
\alias{pct_error}
\title{Percent Error}
\usage{
pct_error(x, y)
}
\arguments{
\item{x}{\link[=vector]{numeric()} observed values}

\item{y}{\link[=vector]{numeric()} forecast values}
}
\value{
\link[=vector]{numeric()} percent error
}
\description{
Percent Error
}
\examples{
pct_error(0:10, 10:0)

}
