% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{pe}
\alias{pe}
\title{Percentage Error}
\usage{
pe(a, f)
}
\arguments{
\item{a}{\code{\link[=numeric]{numeric()}} \code{\link[=vector]{vector()}} of actual values}

\item{f}{\code{\link[=numeric]{numeric()}} \code{\link[=vector]{vector()}} of forecast values}
}
\value{
\code{\link[=numeric]{numeric()}} symmetric percentage error
}
\description{
Percentage Error
}
\examples{
pe(0:10, 10:0)

}
