% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{raw_error}
\alias{raw_error}
\title{Raw Error}
\usage{
raw_error(x, y)
}
\arguments{
\item{x}{\link[=vector]{numeric()} observed values}

\item{y}{\link[=vector]{numeric()} forecast values}
}
\value{
\link[=vector]{numeric()} raw error
}
\description{
Raw Error
}
