% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{re}
\alias{re}
\title{Raw Error}
\usage{
re(a, f)
}
\arguments{
\item{a}{\code{\link[=numeric]{numeric()}} \code{\link[=vector]{vector()}} of actual values}

\item{f}{\code{\link[=numeric]{numeric()}} \code{\link[=vector]{vector()}} of forecast values}
}
\value{
\code{\link[=numeric]{numeric()}} raw error
}
\description{
Raw Error
}
\examples{
re(1:10, 10:1)

}
