% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{rho_or_na}
\alias{rho_or_na}
\title{Pearson Correlation Or NA}
\usage{
rho_or_na(x, max_na = Inf)
}
\arguments{
\item{x}{\code{\link[=numeric]{numeric()}} \code{\link[=matrix]{matrix()}} with \code{ncol = 2}}

\item{max_na}{\code{\link[=integer]{integer()}} scalar}
}
\value{
\code{\link[=numeric]{numeric()}} scalar or \code{NA_real_}
}
\description{
Pearson Correlation Or NA
}
\examples{
# Ends in NA
a <- c(1:9, NA)
b <- c(1:10)
x <- matrix(c(a, b), ncol = 2)
rho_or_na(x)

# Too many NA values
a <- c(1:3, NA, NA, NA, 7:10)
b <- c(1:10)
x <- matrix(c(a, b), ncol = 2)
rho_or_na(x, max_na = 2)

# Correlation coefficient rho
a <- 1:10
b <- c(1:10)
x <- matrix(c(a, b), ncol = 2)
rho_or_na(x)

}
