% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{ricker}
\alias{ricker}
\title{Fit A Ricker Model Via CmdStanR}
\usage{
ricker(
  data,
  chains = 1,
  step_size = 0.1,
  iter_warmup = 250,
  iter_sampling = 750,
  ...
)
}
\arguments{
\item{data}{\code{\link[=list]{list()}}}

\item{chains}{\code{\link[=integer]{integer()}} number of chains}

\item{step_size}{\code{\link[=integer]{integer()}} initial step size}

\item{iter_warmup}{\code{\link[=integer]{integer()}} number of warmup iterations}

\item{iter_sampling}{\code{\link[=integer]{integer()}} number of sampling iterations}

\item{...}{additional arguments to pass to \code{$sample()} method}
}
\value{
\code{\link[=ricker]{ricker()}}
}
\description{
Fit A Ricker Model Via CmdStanR
}
