% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{rms_or_na}
\alias{rms_or_na}
\title{Root Mean Square Or NA}
\usage{
rms_or_na(x, max_na = Inf)
}
\arguments{
\item{x}{\code{\link[=numeric]{numeric()}} \code{\link[=vector]{vector()}}}

\item{max_na}{\code{\link[=integer]{integer()}} scalar}
}
\value{
\code{\link[=numeric]{numeric()}} scalar or \code{NA_real_}
}
\description{
Root Mean Square Or NA
}
\examples{
# Ends in NA
x <- c(1:9, NA)
rms_or_na(x)

# Too many NA values
x <- c(1:3, NA, NA, NA, 7:10)
rms_or_na(x, 2)

# Root mean squared
x <- 1:10
rms_or_na(x)

}
