% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{simulate_environmental_covariates}
\alias{simulate_environmental_covariates}
\title{Simulate Environmental Covariates}
\usage{
simulate_environmental_covariates(
  n,
  mu = 0,
  phi = 0,
  sigma = 1,
  var_names = NULL
)
}
\arguments{
\item{n}{\code{\link[=integer]{integer()}} number of time steps}

\item{mu}{\link[=vector]{numeric()} means}

\item{phi}{\link[=vector]{numeric()}}

\item{sigma}{\link[=vector]{numeric()}}

\item{var_names}{\code{\link[=character]{character()}}}
}
\value{
\code{\link[tibble:tibble]{tibble::tibble()}}
}
\description{
Simulate Environmental Covariates
}
\examples{

simulate_environmental_covariates(15, rep(0, 5), 0.5, 1)

}
