% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simulation}
\alias{simulation}
\title{Simulated Larking Population Dynamics}
\format{
A data frame with 1000 rows and 4 variables:
\describe{
\item{R_3}{abundance of age-3 recruitment}
\item{R_4}{abundance of age-4 recruitment}
\item{R_5}{abundance of age-5 recruitment}
\item{S}{abundance of spawners}
...
}
}
\source{
data-raw/simulation.R
}
\usage{
simulation
}
\description{
A dataset of simulated Larking population dynamics
}
\keyword{datasets}
