% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{smaape}
\alias{smaape}
\title{Scaled Mean Arctangent Absolute Percentage Error}
\usage{
smaape(x, y)
}
\arguments{
\item{x}{\code{\link[=numeric]{numeric()}} \code{\link[=vector]{vector()}} actual values}

\item{y}{\code{\link[=numeric]{numeric()}} \code{\link[=vector]{vector()}} forecast values}
}
\value{
\code{\link[=numeric]{numeric()}} scaled mean arctangent absolute percentage error
}
\description{
Scaled Mean Arctangent Absolute Percentage Error
}
\examples{
x <- rnorm(100, 0, 1)
y1 <- x
smaape(x, y1)

y2 <- rep(mean(x), length(x))
smaape(x, y2)

smaape(c(x, NA), c(y1, NA))

}
\references{
Kim and Kim 2016 doi.org/10.1016/j.ijforecast.2015.12.003
}
