% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{chilko}
\alias{chilko}
\title{Chilko Sockeye Salmon Abundance Data}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 68 rows and 23 columns.
}
\usage{
chilko
}
\description{
Chilko Sockeye Salmon Abundance Data
}
\keyword{datasets}
