% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{mae}
\alias{mae}
\title{Mean Absolute Error}
\usage{
mae(x, y)
}
\arguments{
\item{x}{\code{\link[=numeric]{numeric()}} \code{\link[=vector]{vector()}} actual values}

\item{y}{\code{\link[=numeric]{numeric()}} \code{\link[=vector]{vector()}} forecast values}
}
\value{
\code{\link[=numeric]{numeric()}} mean absolute error
}
\description{
Mean Absolute Error
}
\examples{
mae(1:10, 10:1)

}
