% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{mre}
\alias{mre}
\title{Running Or Total Mean Raw Error}
\usage{
mre(x, y = NULL, running = FALSE)
}
\arguments{
\item{x}{\code{\link[=matrix]{matrix()}} with two columns or \link[=vector]{numeric()}}

\item{y}{\link[=vector]{numeric()}}

\item{running}{\code{\link[=logical]{logical()}}}
}
\value{
\code{\link[=numeric]{numeric()}}
}
\description{
Running Or Total Mean Raw Error
}
\examples{

# Without NA
mre(1:10, 0:9, FALSE)
mre(1:10, 0:9, TRUE)
mre(matrix(c(1:10, 0:9), ncol = 2), running = FALSE)
mre(matrix(c(1:10, 0:9), ncol = 2), running = TRUE)

# With NA
mre(matrix(c(1:10, NA, 1:9), ncol = 2), running = FALSE)
mre(matrix(c(1:10, NA, 1:9), ncol = 2), running = TRUE)

}
