% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{rmse}
\alias{rmse}
\title{Running Or Total Square Root Of The Mean Squared Error}
\usage{
rmse(x, y = NULL, running = FALSE)
}
\arguments{
\item{x}{\code{\link[=matrix]{matrix()}} with two columns or \link[=vector]{numeric()}}

\item{y}{\link[=vector]{numeric()}}

\item{running}{\code{\link[=logical]{logical()}}}
}
\value{
\code{\link[=numeric]{numeric()}}
}
\description{
Running Or Total Square Root Of The Mean Squared Error
}
\examples{

# Without NA
rmse(1:10, 0:9, FALSE)
rmse(1:10, 0:9, TRUE)
rmse(matrix(c(1:10, 0:9), ncol = 2), running = FALSE)
rmse(matrix(c(1:10, 0:9), ncol = 2), running = TRUE)

# With NA
rmse(matrix(c(1:10, NA, 1:9), ncol = 2), running = FALSE)
rmse(matrix(c(1:10, NA, 1:9), ncol = 2), running = TRUE)

}
