% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{sim}
\alias{sim}
\title{Simulate Sockeye Dynamics From A Larkin Stock-Recruitment Model}
\usage{
sim(
  a = 2,
  b = c(6, 2, 1, 1),
  init = rep(1, 8),
  h = 0.2,
  p_bar = c(0.003, 0.917, 0.08),
  sigma = 0.8,
  burn = 100,
  span = 1000,
  seed = NULL
)
}
\arguments{
\item{a}{\code{\link[=numeric]{numeric()}} per-capita log population growth rate at low abundance}

\item{b}{\code{\link[=numeric]{numeric()}} \code{\link[=vector]{vector()}} density dependence parameters}

\item{init}{\code{\link[=numeric]{numeric()}} \code{\link[=vector]{vector()}} initial spawner abundances}

\item{h}{\code{\link[=numeric]{numeric()}} constant harvest rate}

\item{p_bar}{\code{\link[=numeric]{numeric()}} \code{\link[=vector]{vector()}} mean age at maturity proportions}

\item{sigma}{\code{\link[=numeric]{numeric()}} process error standard deviation parameter}

\item{burn}{\code{\link[=integer]{integer()}} number of burn-in steps before the simulation}

\item{span}{\code{\link[=integer]{integer()}} number of steps in the simulation}

\item{seed}{\code{\link[=integer]{integer()}} passed internally to \code{\link[=set.seed]{set.seed()}}}
}
\value{
\code{\link[=data.frame]{data.frame()}} simulated spawner and recruitment at age abundances
}
\description{
Simulate Sockeye Dynamics From A Larkin Stock-Recruitment Model
}
\examples{
s1 <- sim()

}
