% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{srmse}
\alias{srmse}
\title{Scaled Square Root of the Mean Squared Error (SRMSE)}
\usage{
srmse(x, y = NULL, s, running = FALSE)
}
\arguments{
\item{x}{\code{\link[=numeric]{numeric()}} \code{\link[=vector]{vector()}} Observed values}

\item{y}{\code{\link[=numeric]{numeric()}} \code{\link[=vector]{vector()}} Forecast values}

\item{s}{\code{\link[=numeric]{numeric()}} \code{\link[=vector]{vector()}} Forecast values used to scale}

\item{running}{\code{\link[=logical]{logical()}}}
}
\value{
\code{\link[=numeric]{numeric()}}
}
\description{
Scaled Square Root of the Mean Squared Error (SRMSE)
}
\examples{
# Perfect
srmse(0:10, 0:10, 0:10) # NaN because 0/0

# Correlated
srmse(0:10, 1:11, 1:11) # 1 because y == s

# Running
srmse(0:10, 1:11, 1:11, TRUE) # rep(1, 11) because y == s

}
